backend/pyproject.toml
[project]
name = "backend"
version = "1.0.0"
description = "Code Documentation Assistant Backend - FastAPI + Temporal + LangGraph"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Code Doc Assistant Team"}
]
keywords = ["rag", "code-analysis", "documentation", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
]

dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.21",
    "pydantic>=2.12.5",
    "prometheus-client>=0.24.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.22.0",
    "langgraph>=1.0.6",
    "langchain>=1.2.4",
    "langchain-anthropic>=1.3.1",
    "aiohttp>=3.13.3",
    "tiktoken>=0.12.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "structlog>=25.5.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "redis>=5.0.0",
    "tree-sitter==0.21.3",
    "tree-sitter-languages>=1.10.2",
    "chromadb>=1.4.1",
    "anthropic>=0.76.0",
    "openai>=2.15.0",
    "httpx>=0.28.1",
    "limits>=5.6.0",
    "slowapi>=0.1.9",
    "temporalio>=1.21.1",
    "gitpython>=3.1.46",
    "pillow>=12.1.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    # Linting & Formatting
    "ruff>=0.14.11",
    "black>=25.12.0",
    # Pre-commit
    "pre-commit>=4.5.1",
]

[project.urls]
Homepage = "https://github.com/yourusername/code-doc-assistant"
Repository = "https://github.com/yourusername/code-doc-assistant"
Documentation = "https://docs.code-doc-assistant.com"

[tool.ruff]
target-version = "py312"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "W191",   # indentation contains tabs
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["app", "shared"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]
addopts = [
    "--strict-markers",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
